# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: IronClaw Control-Plane API
  version: 0.1.0
  summary: Gateway change-approval, audit, and registry administration for a self-hosted IronClaw control-plane.
  description: |
    The control-plane exposes three surfaces:

    - **Gateway** (`/v1/changes`, `/v1/audit`) — submit capability changes, list
      what is pending, record human approve/reject decisions, and read the
      append-only audit log. Every control-plane mutation (persona, enabled
      tools, packages, wiring, permissions, mounts, create_agent) flows through
      here; there is no bypass path.
    - **Registry** (`/v1/registry/...`) — administer agent groups, messaging
      groups, wirings, users, roles, members, destinations, and read sessions.
      These endpoints return `503` until a registry backend is attached.
    - **Operational probes** (`/healthz`, `/readyz`, `/metrics`).

    All `POST`/`PUT` bodies are JSON. The Go JSON decoder is case-insensitive on
    keys; the property names below are the canonical forms the server **emits**.
  license:
    name: MIT
    url: https://github.com/nivardsec/ironclaw/blob/main/LICENSE
  contact:
    name: IronClaw
    url: https://github.com/nivardsec/ironclaw
...
servers:
  - url: http://127.0.0.1:8787
    description: Local development (`--dev`, loopback bind).
  - url: http://{tailnetHost}:{port}
    description: Production — bound to the control-plane's Tailscale interface address (no public port).
    variables:
...
      port:
...

...
security:
  - bearerAuth: []
...
paths:
...
  /v1/changes:
    post:
      tags: [gateway]
      summary: Submit a capability change.
      description: |
        Submits a `ChangeRequest` to the gateway. The gateway runs its
        deterministic verifier chain and then holds the change for a human
        decision, so submission returns immediately with the assigned id; the
        change is not applied until approved via the decision endpoint. A `null`
        or omitted `ID`/`CreatedAt` is assigned server-side.
      operationId: submitChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRequest"
      responses:
        "202":
          description: Accepted and held pending a decision.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/ChangeID"
                required: [id]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/changes/pending:
    get:
      tags: [gateway]
      summary: List changes awaiting a decision.
      operationId: listPendingChanges
      responses:
        "200":
          description: The pending change requests (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/changes/history:
    get:
      tags: [gateway]
      summary: List applied and rejected changes.
      description: Returns an empty list when no durable history provider is attached (in-memory dev mode).
      operationId: listChangeHistory
      responses:
        "200":
          description: The change history (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/changes/{id}/decision:
    post:
      tags: [gateway]
      summary: Record an approve/reject decision for a pending change.
      operationId: decideChange
      parameters:
        - $ref: "#/components/parameters/ChangeIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionRequest"
      responses:
        "200":
          description: Decision recorded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: recorded
                required: [status]
        "400":
          description: Missing id, invalid JSON, or `outcome` not `approve`/`reject`.
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The change is not pending (already decided, unknown, or held by a conflicting state).
          content:
            text/plain:
              schema:
                type: string
  /v1/audit:
    get:
      tags: [gateway]
      summary: Read recent audit entries.
      description: Returns an empty list when no audit path is attached. Entries are append-only JSONL records.
      operationId: readAudit
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum entries to return (most recent first). Defaults to 100; non-positive values are ignored.
          schema:
            type: integer
            minimum: 1
            default: 100
      responses:
        "200":
          description: Recent audit entries (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token, supplied to the server via `IRONCLAW_API_TOKEN`
        (or `--token` for `ironctl`). It is **all-or-nothing** — there are no
        per-endpoint scopes; a valid token authorizes every non-probe endpoint.
        The token is compared in constant time. `/healthz` and `/readyz` are
        always exempt. When the server is started without a token the check is
        disabled and the mesh (Tailscale) boundary is the sole control.
  parameters:
...
    ChangeIDPath:
      name: id
      in: path
      required: true
      description: The change id.
      schema:
        $ref: "#/components/schemas/ChangeID"
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when the server is started with a token).
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    BadRequest:
      description: Malformed JSON body or invalid field values.
      content:
        text/plain:
          schema:
            type: string
...
    InternalError:
      description: Internal error reading from the backing store.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
...
    ChangeID:
      type: string
      description: Gateway change id (server-assigned `chg_<timestamp>` when omitted).
      examples: ["chg_20260617120000.000000000"]
    UserID:
      type: string
      description: Platform identity, formatted `<channel>:<handle>`.
      examples: ["slack:U123", "telegram:42"]
    AgentGroupID:
      type: string
...
    ChangeKind:
      type: string
      description: The control-plane mutation kind.
      enum:
        - persona
        - enabled_tools
        - packages
        - wiring
        - permissions
        - mounts
        - create_agent
...
    DecisionOutcome:
      type: string
      enum: [approve, reject]
    AuditStage:
      type: string
      enum: [submit, verdict, decision, apply]

    # --- gateway types (internal/contract/gateway.go, internal/host/gateway) ---
    ChangeRequest:
      type: object
      description: |
        The mandatory-mutation protocol payload. `Before`/`After` are
        canonicalized JSON (sorted keys) describing the prior and proposed state
        for the change kind. `ID` and `CreatedAt` are assigned server-side when
        omitted. Field names are the canonical (Go) capitalized forms; the
        decoder also accepts lower-case keys.
      properties:
        ID:
          $ref: "#/components/schemas/ChangeID"
        Kind:
          $ref: "#/components/schemas/ChangeKind"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
        RequestedBy:
          $ref: "#/components/schemas/UserID"
        Before:
          type: object
          additionalProperties: true
          nullable: true
          description: Canonicalized prior state (may be null/empty for a create).
        After:
          type: object
          additionalProperties: true
          nullable: true
          description: Canonicalized proposed state.
        CreatedAt:
          type: string
          format: date-time
      required: [Kind]
    Decision:
      type: object
      description: The recorded outcome of the approval step (as embedded in history).
      properties:
        Outcome:
          $ref: "#/components/schemas/DecisionOutcome"
        DecidedBy:
          $ref: "#/components/schemas/UserID"
        DecidedAt:
          type: string
          format: date-time
      required: [Outcome, DecidedBy, DecidedAt]
    DecisionRequest:
      type: object
      description: Body of the decision endpoint (note the lower-case JSON keys).
      properties:
        outcome:
          $ref: "#/components/schemas/DecisionOutcome"
        decidedBy:
          $ref: "#/components/schemas/UserID"
      required: [outcome, decidedBy]
    HistoryEntry:
      type: object
      description: One applied/rejected change with its recorded decision.
      properties:
        request:
          $ref: "#/components/schemas/ChangeRequest"
        status:
          type: string
          description: Terminal status of the change (e.g. `applied`, `rejected`).
        decision:
          $ref: "#/components/schemas/Decision"
      required: [request, status]
    AuditEntry:
      type: object
      description: One append-only audit record (emitted as JSONL on disk).
      properties:
        time:
          type: string
          format: date-time
        stage:
          $ref: "#/components/schemas/AuditStage"
        changeId:
          $ref: "#/components/schemas/ChangeID"
        kind:
          $ref: "#/components/schemas/ChangeKind"
        detail:
          type: string
      required: [time, stage, changeId]

    # --- registry types (internal/host/registry/registry.go; no JSON tags → capitalized keys) ---
...